Add NullableIndex test. Remove unnecessary static fields in testing/index.
The static variables are actually harmful. In the test there are
some methods calling reverse() of the static array. For Lovefield,
Closure test will reinstantiate the class so that this problem is
not exposed. In Mocha/TypeScript, the array is only instantiated
once (which is the correct behavior), and this flaw is exposed.