Implement EndToEndTransaction test.
There is a problem with attach_Error_Rollback test case.
The original code uses thenCatch(), which is not provided in ES6.
Either async function or plain Promise still result in unhandled
promise rejection, not sure if that's a TypeScript problem or not.

See https://github.com/arthurhsu/lovefield-ts/issues/12 for followup.