Fix the unhandled promise rejection issue in EndToEndTransactionTest.
The original design was to reject the execInternal() promise with a
cancellation error. Inside Closure, the code actually calls resolve_()
and silently scheduleCallback_().
https://github.com/google/closure-library/blob/master/closure/goog/promise/promise.js#L361

This is effectively as resolve(). So the fix here is to just resolve
the issued promise, but throws an error causing the error handling
mechanism inside JS engine to properly reject the promise chain.

Fix: #12