Parse and fix GTS warnings.
This is a big pass of parsing GTS warnings.
At the time of commit, 3 are left as they might be real bugs.

All other delibratedly kept ones are papered out using
// eslint-disable-next-line @typescript-eslint/no-unused-vars

This is because eslint aggressively attempted to remove vars,
even if it's from function signature. For example:

```javascript
interface IBase {
  methodA(var1: type1, var2: type2)
}

class Impl implements IBase {
  methodA(var1: type1, var2: type2) {
    // if vars in arguments are not used, then eslint triggers
  }
}
```

In Lovefield, we attempt to keep all functions' signature consistent
across the board, whether we use it or not.